Elide repeat formatting of static log level
This commit adjusts the `SimpleLogger` to avoid repeatedly formatting the static
log level name. This logger now only formats the record arguments, relying
instead on a match to elide the level format. The string that must be allocated
is now slightly smaller.

We also prefer the use of `write!` and `writeln!` to `print!` and `println!`
allowing this commit to lock stdout, meaning the now two distinct writes do not
have to compete for the stdout lock. One concern here is whether this logger is
called from sync and async code. It's unclear to me if log messages enter a
queue -- meaning that async code is not competing for the lock -- or whether
that ultimately matters.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>