Move `main.rs` into a binary path
This commit moves the `main.rs` down into `src/bin/bottlecap` allowing for code
solely used by `bottlecap` binary to be migrated into that directory, for
additional binaries to be created in the project on the same pattern and for the
`bottlecap` crate to be built indepdently of its binary, useful for micro
benchmarks and the like.

If it becomes desirable we can make a workspace with multiple sub-crates, in the
manner of https://github.com/DataDog/single-machine-performance or
https://github.com/DataDog/saluki/. I don't know that this is necessary right
now.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>