Use jemalloc as global allocator
I imagine that the libc allocator in a lambda environment is glibc. While it's
improved over time jemalloc is often a better all-around allocator. This commit
also demonstrates how one sets a global allocator for a Rust project, park it in
the `main.rs` of your binary and everything linked into that binary -- except
for C code, as the allocator has a prefix and doesn't overwrite `malloc` et al
-- uses the allocator.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>