Add clippy lints to the project, address some dings  (#235)
* Add clippy lints to the project, address some dings

This commit is a follow-on from a suggestion I had early on to make aggressive
use of clippy. The changes here are mostly mechanical however some needless
clones, branches have been caught and fixed. I have left some of the precission
loss dings in place, note that lib.rs has a section of lints explicitly
allowed. These should be an hour or two to grind out.

Assumes that https://github.com/DataDog/datadog-lambda-extension/pull/227 will
merge. Should be rebased once that PR is in `main`.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* Apply suggestions from code review

Co-authored-by: jordan gonzález <30836115+duncanista@users.noreply.github.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>
Co-authored-by: jordan gonzález <30836115+duncanista@users.noreply.github.com>