chore: remove mutex lock for `ContextBuffer`  (#630)
# What?

Removes the Mutex Lock on `ContextBuffer`

# Motivation

We were using `ContextBuffer` to share `tracer_span` so that we can
enrich the top level span generated by the Extension. Instead, we now
share the already existing `InvocationProcessor` lock which allows us to
set the `tracer_span` on the given context.

# Tests

Works as before.

# Notes

There is a known issue which only makes it work on cold start spans,
this was also something that appeared on the previous approach.