chore: remove global state in `InvocationProcessor` (#636)
# What?

Removes global state in `InvocationProcessor` in order to avoid flaky
parenting in inferred spans.

# Motivation

Inferred span parenting became flaky.

Caused by a race condition in which the Universal Instrumentation events
(`/start-invocation` and `/end-invocation`) were happening before we
could process the Telemetry API events.

In an ideal world, the event processing timeline goes as follows:

```
Invoke (I) -> InvocationStart (IS) -> InvocationEnd (IE) -> PlatformRuntimeDone (E)
```

Nonetheless, due to how fast we process events, some might arrive at a
different time. Since these systems send events to the same processor
and they are not synchronized, we now pair each I to an IS, and every IE
to an E.

This is a problem because IS and IE don't send the `request_id`, making
it hard for our algorithm to know to which invocation they belong to, in
the past, we would assume order would prevail, therefore it was OK to
guess that the current data went directly to the global state.
Nonetheless, since other events might arrive, this would rewrite global
state.

This new event handling permits these events to come in a combination of
orders without affecting telemetry data.

# How

1. Removed global state and passed down tracing data directly to the
`ContextBuffer`.
2. Created an unordered events buffer to hold incoming events that might
need a pair in order to process telemetry correctly.

# Notes

1. Also refactored some other minor methods to remove borrows of mutable
data where it wasn't needed (`get_metadata_from_value`).
2. Had to remove a method because couldn't find a way to remove the
borrow mutable problem there (`update_span_context_from_headers`).
3. We still guess the invocation for `InitStart` and `InitReport`
events, because the first time we get these events, a context would have
already been created, so it is safe.