fix: Correctly get oldest invocation in `should_adapt_to_periodic`  (#662)
### What
We switch to periodic flushing if there are at least 20 invocations and
the average time between each invocation is less than 120 seconds. (see
[datadog-agent
implementation](https://github.com/DataDog/datadog-agent/blob/main/pkg/serverless/daemon/interval.go#L47-L76))

In `should_adapt_to_periodic`, we calculate the average time between
invocations (for the last 20 invocations) by taking the time between the
oldest and newest invocations in the buffer, then dividing by 20.

Prior to this, we were assuming the oldest invocation in the buffer was
the last item in the array. This is incorrect because the buffer is a
circular buffer; we can instead use `buffer[head]` to get the oldest
item in the array.

I also added a small improvement where we exit early if the buffer is
not full (meaning <20 invocations have occurred so far).

### Motivation
https://github.com/DataDog/datadog-lambda-extension/issues/652
https://datadoghq.atlassian.net/browse/SVLS-6684