aj/fix route tag (#661)
This is a breaking change, and a feature improvement/enhancement.

It and adds a parameterized path parser for APIGW rest/http events.
Specifically this is most helpful for proxy routes.

Initial versions of the Go extension would just set the `resource` to
the raw route, inclusive of IDs. This caused cardinality issues and
`resource` limit exhaustion for certain customers as serverless would
eat all of the cardinality available.

With the introduction of the next-generation lambda extension, we
changed this to match the implementation of node/python, which tags all
resources for `proxy` routes with `{proxy+}`. That's less helpful for
APM customers doing internal routing with frameworks (but solved the
cardinality issues).

For universal-instrumentation languages, we're unable to rely on the
internal framework route matching so this change looks for:
- integer IDs between REST routes
- UUIDs
- GUIDs
- ULIDs

then replaces the content of the identifier with the previous string
value in the REST URL, so:
```
/dev/proxy_route/users/1234/friends/abdx-3230234-asdfalksdj-2342/pizzas/7
->
/dev/proxy_route/users/{user_id}/friends/{friend_id}/pizzas/{pizza_id}
```

This does diverge from node and python (which we will hopefully extend
this functionality to), but it does strike a useful balance.

This impacts users enabling APIGW Parameterized Routes as well.
Previously we'd use the `resource` argument there, which may use
customer placeholders. Now we override that.

I'm ammenable to changing that (we'd have to check if the `resource`
contains `proxy` and then trigger this behavior instead. But, after spot
checking a few internal users I think this change would result in
basically the same thing. It's still breaking for users with monitors
configured, but that can't be avoided

Fixes #601