fix: cold start tag (#685)
Fixes #682

In #636 we changed this method and inserted the invocation into the
context buffer before setting the init tags:
[link](https://github.com/DataDog/datadog-lambda-extension/commit/94cebb9037c28b17c086c049adb04d77db250a5b#diff-689befe07a6fb7b65e3a7ed5c8949970e0fbbc74e0356f56ea8d829159ae5f14R119).

Fix:

![image](https://github.com/user-attachments/assets/443c6776-b221-4247-b08e-70eefd802625)

This caused the cold start logic to always return false, as it's looking
for that buffer to be empty. A safer implementation could be to check if
the length is 1 and the current request ID matches the item in the
buffer, but this works.