Combine handle_next_invocation and next_event. Don't block for PlatformReport on timeout shutdowns (#692)
1. Refactors handle_next_invocation to call next_event. The methods are
separate because we call next_event in an idle loop, but otherwise we
always call them together from main, so this simplifies things
2. In the shutdown loop, only block for the report line if the shutdown
isn't a timeout. On timeouts, we won't get a report log.

That's not a problem for most use cases where the sandbox will
re-initailize, the telemetry API will re-pass events, and then we can
forward events on the next invocation. But if a function continuosly
times out, we may not forward the custom `task timed out` log until a
few invocations down the line.

~I'll verify if this is true with OOMs~

OOMs can be variable, so i've added a new line to the END log to explain
the status reason. For node it's a runtimeExit:
<img width="941" alt="image"
src="https://github.com/user-attachments/assets/e22a56e1-9276-488d-b333-29ec644489bd"
/>