fix: deadlock in main loop (#696)
Fixes an issue I made in #692. Like all concurrency bugs, this one is
insidious.

At a glance combining `next_event` with `handle_next_invocation` makes a
lot of sense. I wish I could blame it on an LLM, but I had the bright
idea this morning as I reviewed the code.

In the flush_at_end and start cases, it works totally fine.
The problem is that in the high-throughput case where we switch to
periodic mode, we had a tokio::select! block polling over multiple
futures.

In #692 with the combined next_event and handle_next_invocation method,
we end up polling over a future which both waited for the response from
/next_event but then immediately blocked until it could lock the
invocation processor and insert the new request.

Unfortunately this caused a deadlock because inside the tokio::select!
block, we also lock the invocation processor when we need to handle some
event bus events.

If the response from next_event returns and tokio enters that branch, it
may not return if it can't get the lock. But if a request task is
inserting data into the invocation processor, we can deadlock and never
poll that branch again.

![image](https://github.com/user-attachments/assets/cb5795cb-02d3-4298-8f4c-7f28eaf7c7b7)

![image](https://github.com/user-attachments/assets/361ae154-d3b2-478e-b731-e8ada2ba4b8a)