feat: set init_type on init duration metric (#700)
Lambda has been well-known to boost vCPU for on-demand invocations,
decreasing the init_duration time because the CPU is not throttled like
it is during normal invocations. This has lead to issues like the [Doom
Loop](https://aaronstuyvenberg.com/posts/lambda-timeout-doom-loop).

One thing to know is that initializations occurring under `provisioned
concurrency` do *not* receive this CPU boost, which leads to longer than
expected reported `initialization_duration` latencies. This PR adds the
`init_type` tag so users can see the difference. `SnapStart` is never
going to be sent, but it's required to satisfy the enum.

Here are the tags being sent for a 128mb function. Note the 11s init
duration! (it had to retry during the invoke phase!):

![image](https://github.com/user-attachments/assets/cb358db2-60a0-4432-b924-6092cb57a692)

Now, note they are equal at 1769mb RAM:

![image](https://github.com/user-attachments/assets/41ad0735-4c3c-4c00-b48c-e1c381c1ccfb)

As per the Lambda
[docs](https://docs.aws.amazon.com/lambda/latest/dg/configuration-memory.html),
at 1769mb RAM your function receives the equivalent of 1vCPU.

This PR will help prevent customers from being confused as they may
experience longer or shorter initialization durations when using
provisioned concurrency, indicating that a faster initialization was
likely `on-demand`, vs the much slower `provisioned-concurrency`
initialization.