feat: use skip missed tick behavior (#708)
By default, tokio uses the `burst` behavior which will fire ticks
instantly in the case that a task blocks the main thread and the tick is
missed. In our case I believe if there is a long/erroneous flush, this
may cause the race interval to fire even when we'd reset it.

[docs](https://docs.rs/tokio/latest/tokio/time/enum.MissedTickBehavior.html)