feat: [SVLS-7017] bump go to 1.24.4 (#720)
Fixing https://github.com/DataDog/datadog-agent/issues/37967 .

This is actually not necessary. The version of go
we actually use to build the script is defined the the `go.work` file in
the agent rather than the go version we install in the dockerfiles. So
we should be okay (though it's probably good to keep the versions
relatively close).