chore(config): separate config from sources (#709)
# What?

Separates the configuration from sources, allowing it to be used in more
use cases.

# How?

Creates new default configuration and separates the environment
variables and YAML sources from the default.

# Why?

Make it easier to track changes in every source, as the field names
might be different to what they are used at the configuration level.

# Notes

I expect to abstract this even more by providing it as a crate which can
have features, that way customers can only use the sources and product
specific fields they need.

---------

Co-authored-by: Aleksandr Pasechnik <aleksandr.pasechnik@datadoghq.com>
Co-authored-by: Florentin Labelle <florentin.labelle@outlook.fr>