Aj/fix shutdown deadlock (#746)
We added this check for the platform.Report item because we'd
occasionally race past it when shutting down, leading to missing metrics
(namely post-runtime duration for the last request). This predominantly
affected hourly/infrequent functions.

But that change introduced a race condition where sometimes Lambda would
respond to /next with the shutdown event at the same time we'd receive
the last invocation's report event. In some cases, Tokio would run the
report event first followed by the response to /next, where then we'd
hit the shutdown event bus loop and wait for a report event that was
never coming.

This change also adds a 300ms timeout in case of a deadlock.

Jira: https://datadoghq.atlassian.net/browse/SLES-2151