fix(proxy): ensure correct headers are set when forwarding request (#768)
# What?

Fixes an issue where proxy would fail because a response could come as a
streamed buffer

# Why?

We were not taking into account that we are collecting the body before
even forwarding the request to the Lambda Runtime, so headers will sent
`as-is`. Causing issues when a streamed buffer arrived to ourselves as
`transfer-encoding: chunked` which will confuse the runtime since we are
proxying it already.

# Notes

Also added a `env` var check where you can enable proxy as Go, through
the `DD_EXPERIMENTAL_PROXY_ENABLED=true`

---------

Co-authored-by: AJ Stuyvenberg <aj.stuyvenberg@datadoghq.com>
Co-authored-by: Romain Marcadier <romain.marcadier@datadoghq.com>