chore(extension): moves code outside of binary code (#830)
# What?

Extension code was living in the `main.rs` for the binary, I don't think
it was ideal to have it there, so I moved it to its own module.

# Why?

Allows us to then have a dedicated extension module as a library. 

# Notes

Might need to move the TelemetryAPI code into this module, as it's
extension related, that would be in another PR tho