fix(aap): lower severity of several warnings
These produce excessive log spam in customer accounts, for conditions they may
not be able to generally address (either because the cause is outside of their
control, or because they decided the trade-off was acceptable).

The alternative would be allowing granuar configuration of log level limits for
each module, which would be a significantly more involved change.

This also reverts to a behavior that is closer to that of the Go version, which
was silently ignoring these conditions for the most part.