fix: Fix deadlock when snap start and using secret for api key (#853)
# Problem
When a Lambda (1) uses snap start, and (2) specifies Datadog API key
using `DD_API_KEY_SECRET_ARN`, the extension will encounter a deadlock.
For a `RwLock`, the extension first gets a read lock:

https://github.com/DataDog/datadog-lambda-extension/blob/daf633dd003447d78261e7c371838b5af21073a1/bottlecap/src/secrets/decrypt.rs#L45
then tries to get a write lock:

https://github.com/DataDog/datadog-lambda-extension/blob/daf633dd003447d78261e7c371838b5af21073a1/bottlecap/src/secrets/decrypt.rs#L65

which never finishes. This causes the function to time out.

This bug was introduced in
https://github.com/DataDog/datadog-lambda-extension/pull/717.

# This PR
Fix this bug by removing the `RwLock` usage. `AwsCredential` is only
created and used once in `resolve_secrets()`, and `resolve_secrets()` is
only called once, so there's no need to protect this struct with a lock.

# Testing
Tested on a Lambda with:
- Python 3.13 runtime
- snap start
- using `DD_API_KEY_SECRET_ARN`

Before:
- The function timed out.
- Data failed to be sent to Datadog.

After:
- The function finished without timeout.
- Data was sent to Datadog successfully.

# Notes
Jira: https://datadoghq.atlassian.net/browse/SLES-2482