fix(bottlecap): run-time dependency on libgcc_s.so (#868)
An unexpected dependency on `libgcc_s.so` is implied by building on
`musl` platforms with the `-Ctarget-feature=-crt-runtime`, which causes
breakage in some customer functions. Instead of disabling the static C
runtime for the entire build, use a `RUSTC_WRAPPER` script to disable it
only specifically for the build-script of `libddwaf-sys` which is where
`bindgen` needs to be able to dynamically load `libclang.so`.

Additionally, present the `libcompiler-rt.builtins` object to the linker
where necessary because `rust` passes `-nodefaultlibs` to the linker,
but some objects in `libddwaf` depend on compiler intrinsics.

Fixes https://github.com/DataDog/datadog-lambda-extension/issues/858