[SVLS-5988] add integration tests for snapstart (#971)
https://datadoghq.atlassian.net/browse/SVLS-5988
https://datadoghq.atlassian.net/browse/SVLS-7836


## Overview
Adding integration tests for SnapStart, no functional changes to
extension code.
1.) Testing that restore spans are created, and we don't use cold_start
spans. This is a follow up to SVLS-7836.
2.) Testing that trace ids are correctly generated for SVLS-5988. If
multiple lambdas are triggered, they should have unique trace ids. This
is a potential issue for SnapStart because `init` is ran before taking
the snapshot. This doesn't currently work for Python runtime so the test
is commented out. Will follow up with APM Serverless team on correcting
this.

## Testing 
* Newly added integration tests pass.